refactor: isolate legacy parser oracle
