+++ vrt/snapshot/document-path-zero-diff-gate.test.ts
@@ -2,8 +2,8 @@ import { existsSync, readFileSync } from "fs";
 import { dirname, join } from "path";
 import { fileURLToPath } from "url";
 
-import { convertPptxToPngViaParserPath } from "../../packages/pptx-glimpse/src/converter.js";
 import { convertPptxToPngViaDocumentPath } from "../../packages/pptx-glimpse/src/experimental-document-renderer.js";
+import { convertPptxToPngViaParserPath } from "../../packages/pptx-glimpse/src/parser-path-oracle.js";
 import { compareImageBuffers } from "../compare-utils.js";
 import {
   DOCUMENT_PATH_VRT_CASES,
